removed non game files
