--- backend/game/menu.js
@@ -1,1 +1,0 @@
-var music;
