--- frontend/index.html
@@ -1,17 +1,0 @@
-<!DOCTYPE html>
-<html lang="es">
-    <head>
-        <meta charset="UTF-8">
-        <meta http-equiv="X-UA-Compatible" content="IE=edge">
-        <meta name="viewport" content="width=device-width, initial-scale=1.0"> 
-        <title>GoNake by Dreinix</title>
-        
-    </head>
-    <body>
-        <h1>Let's start</h1>  
-        <div id = "game" style="align-self: center;"></div>
-        <script src="//cdn.jsdelivr.net/npm/phaser@3.54.0/dist/phaser.js"></script>
-        <script src="/backend/index.js"></script>
-        <script src="https://unpkg.com/vue@next" ></script> 
-    </body>
-</html>
