--- go.mod
@@ -1,5 +1,0 @@
-module github.com/dreinix/gonake
-
-go 1.16
-
-require github.com/go-chi/chi v1.5.4 // indirect; explicit
